requirements.txt
google-generativeai>=0.8.0
python-dotenv>=1.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
flask>=3.0.0
gunicorn>=21.0.0
fastapi>=0.115.0
uvicorn>=0.32.0
